Move future date checking into `general.js`
We're going to check for future dates in an increasingly large number of places so `general` seems to be correct
