+++ acceptance_tests/features/step_definitions/update-journey-smoke.js
@@ -1,19 +1,10 @@
 'use strict';
 
 const config = require('../../../config');
-const moment = require('moment');
 const base = config.baseUrl || 'http://localhost';
 const port = config.port || '8080';
 const request = require('request');
 
-const urlise = (text) => text.toLowerCase().replace(/[^\w ]+/g, '').replace(/ +/g, '-');
-
-const futureDate = (future) => {
-    let times = future.split(' ');
-    let amount = Number(times[0]);
-    let unit = times[1].slice(0,-1); //day/s month/s
-    return moment().add(amount, unit);
-}
 
 const fixture = {"objectId":"56ec2057084d726f6a4611b8","passport":{"givenNames":"Ahmed","surname":"Shehab Malbas Abdulla Alobeidli","passportNumber":"12345678Y","expiryDate":"2018-04-27","issueDate":"2013-04-28","placeOfIssue":"Kuwait","gender":"F","dateOfBirth":"1987-08-12","placeOfBirth":"KUWAIT","countryOfBirth":"KWT","nationality":"ARE","holdOtherNationalities":"No","heldPreviousNationalities":"No"},"payment":{"orderCode":"Free","feeInPence":0,"paid":true,"paymentDate":"2016-03-18 15:35:51"},"contactDetails":{"emailAddress":"test@gmail.com","secondEmail":"testing@gmail.com","homeAddress":["My House","STREET 32","HOUSE NUMBER 16","","KWT","00000"],"mobilePhoneNumber":"+123-99028325","phoneNumber":"+123-44422555","countryAppliedFrom":"Kuwait","countryAppliedFromCode":"KWT","areYouEmployed":"No"},"journey":{"travelBy":"Plane","arrivalTravel":"BA0156","arrivalDate":"2016-08-21","departureForUKDate":"2016-04-30 08:20:00","arrivalTime":"12:15","portOfArrival":"London - Heathrow","portOfArrivalCode":"LHR","ukAddress":["The hotel in London","Westminster, London W1 7YT,","","","W1 7YT"],"departureDate":"2016-04-21","departureTravel":"BA157","portOfDeparture" : "London - Heathrow, LHR","inwardDepartureCountry":"KWT","inwardDeparturePort":"Kuwait - Kuwait Intl","inwardDeparturePortCode":"KWI","reasonForTravel":"Tourism","travelWithOthers":"No","knowDepartureDetails":"Yes","ukVisitPhoneNumber":"+12345234234","visitMoreThanOnce":"No"},"miscellaneous":{"onBehalfOfMinor":"No","asAnAgent":"No", "flightDetailsCheck":"Yes"},"passportFileId":"56ec1bd8aee7c384447463f0","applicationReference":"EVW123456"};
 
@@ -47,22 +38,4 @@ module.exports = function () {
         this.url(`${base}:${port}/update-journey-details/how-will-you-arrive?evwNumber=EVW123456&token=74d3e5a47154150525964d90d0a99138a2633b49fbd9f829f9387414c1a458ef`)
         .waitForElementVisible('body', 1000);
     });
-
-    this.Then(/^I enter a date "([^"]*)" in the future into "([^"]*)"$/, function (future, field) {
-        let val = futureDate(future);
-        let target = urlise(field);
-
-        this.setValue('#' + target + '-day', val.day());
-        this.setValue('#' + target + '-month', val.month());
-        this.setValue('#' + target + '-year', val.year());
-    });
-
-    // this.Then(/^the "([^"]*)" should contain a date "([^"]*)" in the future$/, function (field, future) {
-    //     let val = futureDate(future);
-    //     let target = urlise(field);
-    //     this.assert.containsText('#' + target + '-day', val.day());
-    //     this.assert.containsText('#' + target + '-month', val.month());
-    //     this.assert.containsText('#' + target + '-year', val.year());
-    // });
-
 }
